+++ app/src/main/java/com/clloret/days/dagger/modules/BroadcastReceiverModule.java
@@ -2,6 +2,7 @@ package com.clloret.days.dagger.modules;
 
 import com.clloret.days.device.receivers.FirstStartBroadcastReceiver;
 import com.clloret.days.device.receivers.ReminderReceiver;
+import com.clloret.days.tasker.receiver.FireReceiver;
 import dagger.Module;
 import dagger.android.ContributesAndroidInjector;
 
@@ -13,5 +14,8 @@ public abstract class BroadcastReceiverModule {
 
   @ContributesAndroidInjector
   abstract ReminderReceiver bindReminderReceiver();
+
+  @ContributesAndroidInjector
+  abstract FireReceiver bindFireReceiver();
 
 }